+++ extras/docs/CHANGELOG.md
@@ -2,4 +2,7 @@
 
 * Version 0.1.0 March 2026
   * first release
+  * Initial implementation of colour bit conversion functions
+  * Added GUI for selecting input and output formats
+  * Added file saving functionality for generated code